cameramodelestimator: fix translation scaling after estimation
The translation of the camera was normalized to the z tranlation (=1).
We need to fix that in a second step because then the initial guess is
not good enough and we often fail to find a good solution!

See http://users.eecs.northwestern.edu/~yingwu/teaching/EECS432/Notes/camera.pdf
for more information.